refactor permission policies: move NotifyInboxPermissionPolicy and related classes to a new policies.py file
